+++ src/OpenAL/OpenTK.Audio.OpenAL/OpenALLibraryNameContainer.cs
@@ -18,6 +18,12 @@ namespace OpenTK.Audio.OpenAL
     public class OpenALLibraryNameContainer
     {
         /// <summary>
+        /// Overrides any platform detection logic and directly searches for the OpenAL library using the provided path.
+        /// If this is <c>null</c> then no override will happen.
+        /// </summary>
+        public static string OverridePath { get; set; } = null;
+
+        /// <summary>
         /// Gets the library name to use on Windows.
         /// </summary>
         public string Windows => "openal32.dll";
@@ -35,7 +41,7 @@ namespace OpenTK.Audio.OpenAL
         /// <summary>
         /// Gets the library name to use on Android.
         /// </summary>
-        public string Android => Linux;
+        public string Android => "libopenal.so";
 
         /// <summary>
         /// Gets the library name to use on iOS.
@@ -44,16 +50,17 @@ namespace OpenTK.Audio.OpenAL
 
         public string GetLibraryName()
         {
-            if (RuntimeInformation.IsOSPlatform(OSPlatform.Linux))
+            if (OverridePath != null)
             {
-                if (RuntimeInformation.IsOSPlatform(OSPlatform.Create("ANDROID")))
-                {
-                    return Android;
-                }
-                else
-                {
-                    return Linux;
-                }
+                return OverridePath;
+            }
+            else if (RuntimeInformation.IsOSPlatform(OSPlatform.Create("ANDROID")))
+            {
+                return Android;
+            }
+            else if (RuntimeInformation.IsOSPlatform(OSPlatform.Linux))
+            {
+                return Linux;
             }
             else if (RuntimeInformation.IsOSPlatform(OSPlatform.FreeBSD))
             {
@@ -63,16 +70,13 @@ namespace OpenTK.Audio.OpenAL
             {
                 return Windows;
             }
+            else if (RuntimeInformation.IsOSPlatform(OSPlatform.Create("IOS")))
+            {
+                return IOS;
+            }
             else if (RuntimeInformation.IsOSPlatform(OSPlatform.OSX))
             {
-                if (RuntimeInformation.IsOSPlatform(OSPlatform.Create("IOS")))
-                {
-                    return IOS;
-                }
-                else
-                {
-                    return MacOS;
-                }
+                return MacOS;
             }
             else
             {