+++ src/OpenTK.Compute/OpenCL/Enums.cs
@@ -6,16 +6,67 @@ namespace OpenTK.Compute.OpenCL
 
     public enum ContextProperties : int
     {
-        ContextPlatform = 0x1084
+        ContextPlatform = 0x1084,
+        InteropUserSync = 0x1085,
+
+        GlContextKHR = 0x2008,
+        EglDisplayKHR = 0x2009,
+        GlxDisplayKHR = 0x200A,
+        WglHDCKHR = 0x200B,
+        CglShareGroupKHR = 0x200C,
+
+        D3D10DeviceKHR = 0x4014,
+
+        AdapterD3D9KHR = 0x2025,
+        AdapterD3D9EXKHR = 0x2026,
+        AdapterDXVAKHR = 0x2027,
+
+        D3D11DeviceKHR = 0x401D,
+
+        MemoryInitializeKHR = 0x2030,
+        TerminateKHR = 0x2032,
     }
 
+    /// <summary>
+    /// The information that can be queried using <see cref="CL.GetPlatformInfo(CLPlatform, PlatformInfo, out byte[])"/>.
+    /// For more info see: https://registry.khronos.org/OpenCL/specs/3.0-unified/html/OpenCL_API.html#platform-queries-table.
+    /// </summary>
     public enum PlatformInfo : uint
     {
+        /// <summary>
+        /// OpenCL profile string. Returns the profile name supported by the implementation. The profile name returned can be one of the following strings:
+        /// FULL_PROFILE - if the implementation supports the OpenCL specification with no need for extensions.
+        /// EMBEDDED_PROFILE - if the implementation supports the OpenCL embedded profile. The embedded profile for OpenCL is described in https://www.khronos.org/registry/OpenCL/specs/3.0-unified/html/OpenCL_API.html#opencl-embedded-profile.
+        /// </summary>
         Profile = 0x0900,
+
+        /// <summary>
+        /// OpenCL version string. Returns the OpenCL version supported by the implementation. This version string has the following format:
+        /// "OpenCL {major_version.minor_version} {platform-specific information}"
+        /// The major_version.minor_version value returned will be one of 1.0, 1.1, 1.2, 2.0, 2.1, 2.2 or 3.0.
+        /// </summary>
         Version = 0x0901,
+
+        /// <summary>
+        /// Platform name string.
+        /// </summary>
         Name = 0x0902,
+
+        /// <summary>
+        /// Platform vendor string.
+        /// </summary>
         Vendor = 0x0903,
+
+        /// <summary>
+        /// Returns a space separated list of extension names (the extension names themselves do not contain any spaces) supported by the platform. Each extension that is supported by all devices associated with this platform must be reported here.
+        /// </summary>
         Extensions = 0x0904,
+
+        /// <summary>
+        /// Introduced in version 2.1.
+        /// Returns the resolution of the host timer in nanoseconds as used by <see cref="CL.GetHostTimer(CLDevice, IntPtr)"/>.
+        /// This value must be 0 for devices that do not support device and host timer synchronization.
+        /// </summary>
         PlatformHostTimerResolution = 0x0905,
         PlatformIcdSuffix = 0x0920
     }