+++ src/OpenTK.Compute/OpenCL/OpenCLLibraryNameContainer.cs
@@ -5,11 +5,17 @@ namespace OpenTK.Compute.OpenCL
 {
     public class OpenCLLibraryNameContainer
     {
+        /// <summary>
+        /// Overrides any platform detection logic and directly searches for the OpenCL library using the provided path.
+        /// If this is <c>null</c> then no override will happen.
+        /// </summary>
+        public static string OverridePath { get; set; } = null;
+
         public string Linux => "libOpenCL.so.1";
 
         public string MacOS => "/System/Library/Frameworks/OpenCL.framework/OpenCL";
 
-        public string Android => Linux;
+        public string Android => "libOpenCL.so";
 
         public string IOS => MacOS;
 
@@ -17,31 +23,29 @@ namespace OpenTK.Compute.OpenCL
 
         public string GetLibraryName()
         {
-            if (RuntimeInformation.IsOSPlatform(OSPlatform.Linux))
+            if (OverridePath != null)
             {
-                if (RuntimeInformation.IsOSPlatform(OSPlatform.Create("ANDROID")))
-                {
-                    return Android;
-                }
-                else
-                {
-                    return Linux;
-                }
+                return OverridePath;
+            }
+            else if (RuntimeInformation.IsOSPlatform(OSPlatform.Create("ANDROID")))
+            {
+                return Android;
+            }
+            else if (RuntimeInformation.IsOSPlatform(OSPlatform.Linux))
+            {
+                return Linux;
             }
             else if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
             {
                 return Windows;
             }
+            else if (RuntimeInformation.IsOSPlatform(OSPlatform.Create("IOS")))
+            {
+                return IOS;
+            }
             else if (RuntimeInformation.IsOSPlatform(OSPlatform.OSX))
             {
-                if (RuntimeInformation.IsOSPlatform(OSPlatform.Create("IOS")))
-                {
-                    return IOS;
-                }
-                else
-                {
-                    return MacOS;
-                }
+                return MacOS;
             }
             else
             {